Fix tlsOptions compilation

+++ Sources/Pulse/RemoteLogger/RemoteLogger-Connection.swift
@@ -284,7 +284,7 @@ extension NWParameters {
         )
         sec_protocol_options_append_tls_ciphersuite(
             tlsOptions.securityProtocolOptions,
-            tls_ciphersuite_t(rawValue: TLS_PSK_WITH_AES_128_GCM_SHA256)!
+            tls_ciphersuite_t(rawValue: UInt16(TLS_PSK_WITH_AES_128_GCM_SHA256))!
         )
         return tlsOptions
     }
